# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ZakApp API
  description: |
    A privacy-first Islamic Zakat calculator API with comprehensive asset management and yearly tracking capabilities.

    ## Constitutional Principles
    This API adheres to ZakApp's constitutional principles:
    - **Privacy & Security First**: All endpoints implement encryption, authentication, and data protection
    - **Islamic Compliance**: Calculations and methodologies follow authentic Islamic jurisprudence
    - **User-Centric Design**: RESTful design with clear, intuitive endpoints
    - **Quality & Reliability**: Comprehensive validation, error handling, and monitoring
    - **Transparency & Trust**: Detailed documentation and clear response formats
  version: 1.0.0
  contact:
    name: ZakApp Support
    email: support@zakapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.zakapp.com/v1
    description: Production server
  - url: http://localhost:3001/v1
    description: Local development server (API server default)
  - url: http://localhost:5000/api
    description: Local development server (client default)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human readable error message
          example: "Invalid input data"
        details:
          type: array
          items:
            type: string
          description: Additional validation error details
      required:
        - success
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data (varies by endpoint)
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
              example: "2024-09-29T12:00:00Z"
            requestId:
              type: string
              example: "req_abc123"
            version:
              type: string
              example: "1.0.0"
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Primary Savings Account"
        category:
          type: string
          enum: ["cash", "gold", "silver", "crypto", "business", "investment", "real_estate", "debt"]
          example: "cash"
        amount:
          type: number
          minimum: 0
          example: 50000.00
        currency:
          type: string
          example: "USD"
        isZakatable:
          type: boolean
          example: true
        notes:
          type: string
          example: "Emergency savings"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
      required:
        - id
        - userId
        - name
        - category
        - amount
        - currency
        - isZakatable
        - createdAt
        - updatedAt
...
paths:
...
  /assets:
    get:
      summary: Get user assets
      description: Retrieve all assets for the authenticated user
      tags: ["Assets"]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - name: category
          in: query
          schema:
            type: string
            enum: ["cash", "gold", "silver", "crypto", "business", "investment", "real_estate", "debt"]
          description: Filter by asset category
      responses:
        "200":
          description: Assets retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          assets:
                            type: array
                            items:
                              $ref: "#/components/schemas/Asset"
                          pagination:
                            $ref: "#/components/schemas/PaginationMeta"
                        required:
                          - assets
                          - pagination
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create new asset
      description: Create a new asset for the authenticated user
      tags: ["Assets"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "Primary Savings Account"
                category:
                  type: string
                  enum: ["cash", "gold", "silver", "crypto", "business", "investment", "real_estate", "debt"]
                  example: "cash"
                amount:
                  type: number
                  minimum: 0
                  example: 50000.00
                currency:
                  type: string
                  default: "USD"
                  example: "USD"
                isZakatable:
                  type: boolean
                  default: true
                  example: true
                notes:
                  type: string
                  maxLength: 500
                  example: "Emergency savings"
              required:
                - name
                - category
                - amount
      responses:
        "201":
          description: Asset created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /methodologies:
    get:
      summary: Get available Zakat methodologies
      description: Retrieve all supported Zakat calculation methodologies with their configurations
      tags: ["Methodologies"]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Methodologies retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          methodologies:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                  enum: ["standard", "hanafi", "shafi", "maliki", "hanbali"]
                                  example: "standard"
                                name:
                                  type: string
                                  example: "Standard Methodology"
                                description:
                                  type: string
                                  example: "Standard Zakat calculation following majority scholarly opinion"
                                nisabGoldGrams:
                                  type: number
                                  minimum: 0
                                  example: 85.0
                                nisabSilverGrams:
                                  type: number
                                  minimum: 0
                                  example: 595.0
                                zakatRate:
                                  type: number
                                  minimum: 0
                                  maximum: 1
                                  example: 0.025
                                references:
                                  type: array
                                  items:
                                    type: string
                                  example: ["Quran 9:60", "Sahih Muslim 978"]
                              required:
                                - id
                                - name
                                - description
                                - nisabGoldGrams
                                - nisabSilverGrams
                                - zakatRate
                        required:
                          - methodologies
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...